requirements.txt
blinker==1.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cryptowatch-sdk==0.0.14
Flask==1.1.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.10.0
numpy==1.19.5
protobuf==3.14.0
python-dotenv==0.15.0
PyYAML==5.3.1
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.22
urllib3==1.26.2
websocket-client==0.57.0
Werkzeug==1.0.1
